Removed duplicate variables for navigation
