+++ frontend/app/src/main/java/com/example/haengsha/MainActivity.kt
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.example.haengsha.ui.screens.ScreenNavigate
 import com.example.haengsha.ui.theme.HaengshaTheme
 
 class MainActivity : ComponentActivity() {
