--- frontend/app/src/main/java/com/example/haengsha/Routes.kt
@@ -1,9 +1,0 @@
-package com.example.haengsha
-
-sealed class Routes(val route: String) {
-    object SignUp : Routes("SignUp")
-    object ForgotPassword : Routes("ForgotPassword")
-    object Login : Routes("Login")
-    object Dashboard : Routes("Dashboard")
-    object Home : Routes("Home")
-}
