+++ frontend/app/src/main/java/com/example/haengsha/ui/screens/HomeScreen.kt
@@ -7,7 +7,6 @@ import androidx.navigation.compose.rememberNavController
 
 @Composable
 fun HomeScreen(navController: NavHostController) {
-    val navController = rememberNavController()
 
 }
