+++ frontend/app/src/main/java/com/example/haengsha/ui/screens/SignupScreen.kt
@@ -7,7 +7,6 @@ import androidx.navigation.compose.rememberNavController
 
 @Composable
 fun SignupScreen(navController: NavHostController) {
-    val navController = rememberNavController()
 
 }
